Pipeline for bulk long read RNA-seq data processing

Semi-supervised isofrom detection and annotation for long read data. This variant is meant for bulk samples. Specific parameters can be configured in the config file (see create_config), input files are specified via arguments.

Usage

BulkPipeline(
  config_file,
  outdir,
  fastq,
  annotation,
  genome_fa,
  minimap2,
  samtools,
  k8
)

Arguments

config_file

Path to the JSON configuration file. See create_config for creating one.

outdir

Path to the output directory. If it does not exist, it will be created.

fastq

Path to the FASTQ file or a directory containing FASTQ files. Each file will be processed as an individual sample.

annotation

The file path to the annotation file in GFF3 / GTF format.

genome_fa

The file path to the reference genome in FASTA format.

minimap2

(optional) The path to the minimap2 binary. If not provided, FLAMES will use a copy from bioconda via basilisk.

samtools

(optional) The path to the samtools binary. If not provided, FLAMES will use a copy from bioconda via basilisk.

k8

(optional) The path to the k8 binary to run paftools.js from minimap2. If not provided, FLAMES will use a copy from bioconda via basilisk.

Value

A FLAMES.Pipeline object. The pipeline could be run using run_FLAMES, and / or resumed using resume_FLAMES.

Details

By default FLAMES use minimap2 for read alignment. After the genome alignment step (do_genome_align), FLAMES summarizes the alignment for each read by grouping reads with similar splice junctions to get a raw isoform annotation (do_isoform_id). The raw isoform annotation is compared against the reference annotation to correct potential splice site and transcript start/end errors. Transcripts that have similar splice junctions and transcript start/end to the reference transcript are merged with the reference. This process will also collapse isoforms that are likely to be truncated transcripts. If isoform_id_bambu is set to TRUE, bambu::bambu will be used to generate the updated annotations. Next is the read realignment step (do_read_realign), where the sequence of each transcript from the update annotation is extracted, and the reads are realigned to this updated transcript_assembly.fa by minimap2. The transcripts with only a few full-length aligned reads are discarded. The reads are assigned to transcripts based on both alignment score, fractions of reads aligned and transcript coverage. Reads that cannot be uniquely assigned to transcripts or have low transcript coverage are discarded. The UMI transcript count matrix is generated by collapsing the reads with the same UMI in a similar way to what is done for short-read scRNA-seq data, but allowing for an edit distance of up to 2 by default. Most of the parameters, such as the minimal distance to splice site and minimal percentage of transcript coverage can be modified by the JSON configuration file (config_file).

See also

Examples

# download example data
temp_path <- tempfile()
bfc <- BiocFileCache::BiocFileCache(temp_path, ask = FALSE)
file_url <-
  "https://raw.githubusercontent.com/OliverVoogd/FLAMESData/master/data"
fastq1 <- bfc[[names(BiocFileCache::bfcadd(bfc, "Fastq1", paste(file_url, "fastq/sample1.fastq.gz", sep = "/")))]]
fastq2 <- bfc[[names(BiocFileCache::bfcadd(bfc, "Fastq2", paste(file_url, "fastq/sample2.fastq.gz", sep = "/")))]]
annotation <- bfc[[names(BiocFileCache::bfcadd(bfc, "annot.gtf", paste(file_url, "SIRV_isoforms_multi-fasta-annotation_C_170612a.gtf", sep = "/")))]]
genome_fa <- bfc[[names(BiocFileCache::bfcadd(bfc, "genome.fa", paste(file_url, "SIRV_isoforms_multi-fasta_170612a.fasta", sep = "/")))]]
fastq_dir <- paste(temp_path, "fastq_dir", sep = "/") # the downloaded fastq files need to be in a directory to be merged together
dir.create(fastq_dir)
file.copy(c(fastq1, fastq2), fastq_dir)
#> [1] TRUE TRUE
unlink(c(fastq1, fastq2)) # the original files can be deleted
outdir <- tempfile()
dir.create(outdir)

ppl <- BulkPipeline(
  fastq = fastq_dir, annotation = annotation, genome_fa = genome_fa,
  config_file = create_config(outdir, type = "sc_3end", threads = 1, no_flank = TRUE),
  outdir = outdir
)
#> Writing configuration parameters to:  /tmp/Rtmp7MvBFG/file9af33ba80788/config_file_39667.json 
#> Configured steps: 
#> 	genome_alignment: TRUE
#> 	isoform_identification: TRUE
#> 	read_realignment: TRUE
#> 	transcript_quantification: TRUE
ppl <- run_FLAMES(ppl) # run the pipeline
#> Running step: genome_alignment
#> Aligning sample 9af311c3fbf5_sample2 -> /tmp/Rtmp7MvBFG/file9af33ba80788/9af311c3fbf5_sample2_align2genome.bam
#> Sorting BAM files by genome coordinates with 1 threads...
#> Indexing bam files
#> Aligning sample 9af346e2481e_sample1 -> /tmp/Rtmp7MvBFG/file9af33ba80788/9af346e2481e_sample1_align2genome.bam
#> Sorting BAM files by genome coordinates with 1 threads...
#> Indexing bam files
#> Running step: isoform_identification
#> Import genomic features from the file as a GRanges object ... 
#> OK
#> Prepare the 'metadata' data frame ... 
#> OK
#> Make the TxDb object ... 
#> Warning: genome version information is not available for this TxDb object
#> OK
#> Running step: read_realignment
#> Realigning sample 9af311c3fbf5_sample2 -> /tmp/Rtmp7MvBFG/file9af33ba80788/9af311c3fbf5_sample2_realign2transcript.bam
#> Skipped sorting BAM files.
#> Realigning sample 9af346e2481e_sample1 -> /tmp/Rtmp7MvBFG/file9af33ba80788/9af346e2481e_sample1_realign2transcript.bam
#> Skipped sorting BAM files.
#> Running step: transcript_quantification
experiment(ppl) # get the result as SummarizedExperiment
#> class: SummarizedExperiment 
#> dim: 71 2 
#> metadata(0):
#> assays(1): counts
#> rownames(71): SIRV108 SIRV109 ... SIRV706 SIRV7_56032_147957_1
#> rowData names(0):
#> colnames(2): 9af311c3fbf5_sample2 9af346e2481e_sample1
#> colData names(0):