Resume a FLAMES pipeline

This function resumes a FLAMES pipeline by running configured but unfinished steps.

Usage

resume_FLAMES(pipeline)

# S4 method for class 'FLAMES.Pipeline'
resume_FLAMES(pipeline)

Arguments

pipeline

A FLAMES.Pipeline object.

Value

An updated FLAMES.Pipeline object.

See also

Examples

pipeline <- example_pipeline("BulkPipeline")
#> Writing configuration parameters to:  /tmp/RtmpcUsJBX/file347c5c14b84b/config_file_13436.json 
#> Configured steps: 
#> 	genome_alignment: TRUE
#> 	isoform_identification: TRUE
#> 	read_realignment: TRUE
#> 	transcript_quantification: TRUE
#> samtools not found, will use Rsamtools package instead
pipeline <- run_step(pipeline, "genome_alignment")
#> ── Running step: genome_alignment @ Wed Jul  2 03:07:06 2025 ───────────────────
#> Creating junction bed file from GFF3 annotation.
#> Aligning sample sample1 -> /tmp/RtmpcUsJBX/file347c5c14b84b/sample1_align2genome.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Sorting BAM files by genome coordinates with 1 threads...
#> Indexing bam files
#> Aligning sample sample2 -> /tmp/RtmpcUsJBX/file347c5c14b84b/sample2_align2genome.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Sorting BAM files by genome coordinates with 1 threads...
#> Indexing bam files
#> Aligning sample sample3 -> /tmp/RtmpcUsJBX/file347c5c14b84b/sample3_align2genome.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Sorting BAM files by genome coordinates with 1 threads...
#> Indexing bam files
pipeline <- resume_FLAMES(pipeline)
#> Resuming pipeline from step: isoform_identification
#> ── Running step: isoform_identification @ Wed Jul  2 03:07:06 2025 ─────────────
#> Import genomic features from the file as a GRanges object ... 
#> OK
#> Prepare the 'metadata' data frame ... 
#> OK
#> Make the TxDb object ... 
#> Warning: genome version information is not available for this TxDb object
#> OK
#> ── Running step: read_realignment @ Wed Jul  2 03:07:07 2025 ───────────────────
#> Realigning sample sample1 -> /tmp/RtmpcUsJBX/file347c5c14b84b/sample1_realign2transcript.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Skipped sorting BAM files.
#> Realigning sample sample2 -> /tmp/RtmpcUsJBX/file347c5c14b84b/sample2_realign2transcript.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Skipped sorting BAM files.
#> Realigning sample sample3 -> /tmp/RtmpcUsJBX/file347c5c14b84b/sample3_realign2transcript.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Skipped sorting BAM files.
#> ── Running step: transcript_quantification @ Wed Jul  2 03:07:07 2025 ──────────