Execute a FLAMES pipeline

This function runs the FLAMES pipeline. It will run all steps in the pipeline.

Usage

run_FLAMES(pipeline, overwrite = FALSE)

# S4 method for class 'FLAMES.Pipeline'
run_FLAMES(pipeline, overwrite = FALSE)

Arguments

pipeline

A FLAMES.Pipeline object.

overwrite

(optional) If TRUE, the pipeline will be re-run even if some steps are already completed.

Value

An updated FLAMES.Pipeline object.

See also

Examples

pipeline <- example_pipeline("BulkPipeline")
#> Writing configuration parameters to:  /tmp/RtmpXoUSS6/file9b2839e05ad2/config_file_39720.json 
#> Configured steps: 
#> 	genome_alignment: TRUE
#> 	isoform_identification: TRUE
#> 	read_realignment: TRUE
#> 	transcript_quantification: TRUE
#> samtools not found, will use Rsamtools package instead
pipeline <- run_FLAMES(pipeline)
#> ── Running step: genome_alignment @ Fri Sep 12 08:00:03 2025 ───────────────────
#> Creating junction bed file from GFF3 annotation.
#> Aligning sample sample1 -> /tmp/RtmpXoUSS6/file9b2839e05ad2/sample1_align2genome.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Sorting BAM files by genome coordinates with 1 threads...
#> Indexing bam files
#> Aligning sample sample2 -> /tmp/RtmpXoUSS6/file9b2839e05ad2/sample2_align2genome.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Sorting BAM files by genome coordinates with 1 threads...
#> Indexing bam files
#> Aligning sample sample3 -> /tmp/RtmpXoUSS6/file9b2839e05ad2/sample3_align2genome.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Sorting BAM files by genome coordinates with 1 threads...
#> Indexing bam files
#> ── Running step: isoform_identification @ Fri Sep 12 08:00:03 2025 ─────────────
#> ── Running step: read_realignment @ Fri Sep 12 08:00:03 2025 ───────────────────
#> Realigning sample sample1 -> /tmp/RtmpXoUSS6/file9b2839e05ad2/sample1_realign2transcript.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Skipped sorting BAM files.
#> Realigning sample sample2 -> /tmp/RtmpXoUSS6/file9b2839e05ad2/sample2_realign2transcript.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Skipped sorting BAM files.
#> Realigning sample sample3 -> /tmp/RtmpXoUSS6/file9b2839e05ad2/sample3_realign2transcript.bam
#> Warning: samtools not found, using Rsamtools instead, this could be slower and might fail for large BAM files.
#> Skipped sorting BAM files.
#> ── Running step: transcript_quantification @ Fri Sep 12 08:00:04 2025 ──────────