Impute missing transcript counts

Impute missing transcript counts using a shared nearest neighbor graph

Usage

sc_impute_transcript(combined_sce, dimred = "PCA", ...)

Arguments

combined_sce

A SingleCellExperiment object with gene counts and a "transcript" altExp slot.

dimred

The name of the reduced dimension to use for building the shared nearest neighbor graph.

...

Additional arguments to pass to scran::buildSNNGraph. E.g. k = 30.

Value

A SingleCellExperiment object with imputed logcounts assay in the "transcript" altExp slot.

Details

For cells with NA values in the "transcript" altExp slot, this function imputes the missing values from cells with non-missing values. A shared nearest neighbor graph is built using reduced dimensions from the SingleCellExperiment object, and the imputation is done where the imputed value for a cell is the weighted sum of the transcript counts of its neighbors. Imputed values are stored in the "logcounts" assay of the "transcript" altExp slot. The "counts" assay is used to obtain logcounts but left unchanged.

Examples

sce <- SingleCellExperiment::SingleCellExperiment(assays = list(counts = matrix(rpois(50, 5), ncol = 10)))
long_read <- SingleCellExperiment::SingleCellExperiment(assays = list(counts = matrix(rpois(40, 5), ncol = 10)))
SingleCellExperiment::altExp(sce, "transcript") <- long_read
SingleCellExperiment::counts(SingleCellExperiment::altExp(sce))[,1:2] <- NA
SingleCellExperiment::counts(SingleCellExperiment::altExp(sce))
#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10]
#> [1,]   NA   NA    4    5    7    7    4    4    7     4
#> [2,]   NA   NA    3    2    4    4    4    4    2     5
#> [3,]   NA   NA    6    7    9    4    3    3    4     3
#> [4,]   NA   NA    4    6    4    6    9    6    2     5
imputed_sce <- sc_impute_transcript(sce, k = 4)
#> Warning: more singular values/vectors requested than available
#> Warning: You're computing too large a percentage of total singular values, use a standard svd instead.
#> Imputing transcript counts ...
SingleCellExperiment::logcounts(SingleCellExperiment::altExp(imputed_sce))
#> 4 x 10 Matrix of class "dgeMatrix"
#>          [,1]     [,2]     [,3]     [,4]     [,5]     [,6]     [,7]     [,8]
#> [1,] 2.721542 2.630257 2.443919 2.515700 2.701595 2.866249 2.255501 2.443919
#> [2,] 2.065714 2.127489 2.114661 1.529821 2.051662 2.200140 2.255501 2.443919
#> [3,] 2.502580 2.464166 2.937677 2.927185 3.014020 2.200140 1.937815 2.114661
#> [4,] 2.499525 2.588603 2.443919 2.736064 2.051662 2.676461 3.246978 2.937677
#>          [,9]    [,10]
#> [1,] 3.294008 2.443919
#> [2,] 1.814209 2.711819
#> [3,] 2.592955 2.114661
#> [4,] 1.814209 2.711819