BLAZE Assign reads to cell barcodes.

Uses BLAZE to generate barcode list and assign reads to cell barcodes.

Usage

blaze(expect_cells, fq_in, ...)

Arguments

expect_cells

Integer, expected number of cells. Note: this could be just a rough estimate. E.g., the targeted number of cells.

fq_in

File path to the fastq file used as a query sequence file

...

Additional BLAZE configuration parameters. E.g., setting `'output-prefix'='some_prefix'` is equivalent to specifying `–output-prefix some_prefix` in BLAZE; Similarly, `overwrite=TRUE` is equivalent to switch on the `–overwrite` option. Note that the specified parameters will override the parameters specified in the configuration file. All available options can be found at https://github.com/shimlab/BLAZE.

Value

A data.frame summarising the reads aligned. Other outputs are written to disk. The details of the output files can be found at https://github.com/shimlab/BLAZE.

Examples

temp_path <- tempfile()
bfc <- BiocFileCache::BiocFileCache(temp_path, ask = FALSE)
fastq1_url <- 'https://raw.githubusercontent.com/shimlab/BLAZE/main/test/data/FAR20033_pass_51e510db_100.fastq'
fastq1 <- bfc[[names(BiocFileCache::bfcadd(bfc, 'Fastq1', fastq1_url))]]
outdir <- tempfile()
dir.create(outdir)
if (FALSE) { # \dontrun{
blaze(expect_cells=10, fastq1, overwrite=TRUE)
} # }