Create Configuration File From Arguments

Create Configuration File From Arguments

Usage

create_config(outdir, type = "sc_3end", ...)

Arguments

outdir

the destination directory for the configuratio nfile

type

use an example config, available values:

"sc_3end"

- config for 10x 3' end ONT reads

"SIRV"

- config for the SIRV example reads

...

Configuration parameters.

seed

- Integer. Seed for minimap2.

threads

- Number of threads to use.

do_barcode_demultiplex

- Boolean. Specifies whether to run the barcode demultiplexing step.

do_genome_alignment

- Boolean. Specifies whether to run the genome alignment step. TRUE is recommended

do_gene_quantification

- Boolean. Specifies whether to run gene quantification using the genome alignment results. TRUE is recommended

do_isoform_identification

- Boolean. Specifies whether to run the isoform identification step. TRUE is recommended

bambu_isoform_identification

- Boolean. Whether to use Bambu for isoform identification.

multithread_isoform_identification

- Boolean. Whether to use FLAMES' new multithreaded Cpp implementation for isoform identification.

do_read_realignment

- Boolean. Specifies whether to run the read realignment step. TRUE is recommended

do_transcript_quantification

- Boolean. Specifies whether to run the transcript quantification step. TRUE is recommended

barcode_parameters

- List. Parameters for barcode demultiplexing passed to find_barcode (except fastq, barcodes_file, stats_out, reads_out) and threads, which are set by the pipeline, see ?find_barcode for more details.

generate_raw_isoform

- Boolean. Whether to generate all isoforms for debugging purpose.

max_dist

- Maximum distance allowed when merging splicing sites in isoform consensus clustering.

max_ts_dist

- Maximum distance allowed when merging transcript start/end position in isoform consensus clustering.

max_splice_match_dist

- Maximum distance allowed when merging splice site called from the data and the reference annotation.

min_fl_exon_len

- Minimum length for the first exon outside the gene body in reference annotation. This is to correct the alignment artifact

max_site_per_splice

- Maximum transcript start/end site combinations allowed per splice chain

min_sup_cnt

- Minimum number of read support an isoform decrease this number will significantly increase the number of isoform detected.

min_cnt_pct

- Minimum percentage of count for an isoform relative to total count for the same gene.

min_sup_pct

- Minimum percentage of count for an splice chain that support a given transcript start/end site combination.

strand_specific

- 0, 1 or -1. 1 indicates if reads are in the same strand as mRNA, -1 indicates reads are reverse complemented, 0 indicates reads are not strand specific.

remove_incomp_reads

- The strenge of truncated isoform filtering. larger number means more stringent filtering.

use_junctions

- whether to use known splice junctions to help correct the alignment results

no_flank

- Boolean. for synthetic spike-in data. refer to Minimap2 document for detail

use_annotation

- Boolean. whether to use reference to help annotate known isoforms

min_tr_coverage

- Minimum percentage of isoform coverage for a read to be aligned to that isoform

min_read_coverage

- Minimum percentage of read coverage for a read to be uniquely aligned to that isoform

Value

file path to the config file created

Details

Create a list object containing the arguments supplied in a format usable for the FLAMES pipeline. Also writes the object to a JSON file, which is located with the prefix 'config_' in the supplied outdir. Default values from extdata/config_sclr_nanopore_3end.json will be used for unprovided parameters.

Examples

# create the default configuration file
outdir <- tempdir()
config <- create_config(outdir)
#> Writing configuration parameters to:  /tmp/Rtmprnnob2/config_file_7722.json